+++ src/CMakeLists.txt
@@ -6,6 +6,8 @@ add_subdirectory(ipc)
 add_subdirectory(audio)
 add_subdirectory(lib)
 add_subdirectory(math)
+add_subdirectory(module)
+
 if(CONFIG_SAMPLES)
 	add_subdirectory(samples)
 endif()
